The "Excel Data Copy Script" is a Python script that automates the process of copying specific columns from one Excel file to a new Excel file. It utilizes the win32com library to interact with Excel and allows users to select the columns they want to copy. This script is designed to simplify data manipulation tasks by streamlining the process of extracting and saving selected data.
The "Excel Data Copy Script" is a Python utility that simplifies the task of copying data from one Excel file to another. Whether you are working with financial data, research results, or any other type of tabular data, this script can save you time and effort.

Key Features:

User-Friendly: The script prompts the user to enter the names of the columns they want to copy. Columns are specified by their headers, which are separated by commas. This user input is flexible and allows for a customized data extraction process.

Automation: The script leverages the win32com library to interact with Microsoft Excel, which means it can perform the data copying process automatically without requiring manual copying and pasting.

Output Control: The script creates a new Excel file in the same directory as the source file, making it easy to organize and manage your data. The new file is named "Hammadde.xlsx" and contains the selected columns from the source file.

Error Handling: In the event of any errors, the script will display an error message and wait for 10 seconds before exiting, providing feedback on potential issues.

How to Use:

Place the script in the same directory as your source Excel file, which should be named content.xlsx.

Run the script.

You will be prompted to input the names of the columns you want to copy. Specify the column names separated by commas.

The script will proceed to create a new Excel file named "Hammadde.xlsx" in the same directory, containing the selected columns from the source file.

In case of errors, the script will provide an error message and a 10-second delay before exiting.

This script is a valuable tool for anyone working with Excel data who wants to streamline the process of data extraction and save time on manual copying and pasting. Feel free to modify and adapt the script to meet your specific requirements.

Author
Furkan

For any questions, improvements, or assistance, please don't hesitate to contact furkancakir.dev